Moved logic to construct StubRequest from HttpRequest to StubRepository
