+++ main/java/io/github/azagniotov/stubby4j/stubs/StubRepository.java
@@ -12,6 +12,7 @@ import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
@@ -27,8 +28,13 @@ import java.util.concurrent.atomic.AtomicLong;
 import static io.github.azagniotov.stubby4j.stubs.StubResponse.notFoundResponse;
 import static io.github.azagniotov.stubby4j.stubs.StubResponse.redirectResponse;
 import static io.github.azagniotov.stubby4j.stubs.StubResponse.unauthorizedResponse;
+import static io.github.azagniotov.stubby4j.utils.CollectionUtils.constructParamMap;
+import static io.github.azagniotov.stubby4j.utils.HandlerUtils.extractPostRequestBody;
+import static io.github.azagniotov.stubby4j.utils.ObjectUtils.isNotNull;
 import static io.github.azagniotov.stubby4j.utils.ReflectionUtils.injectObjectFields;
+import static io.github.azagniotov.stubby4j.utils.StringUtils.toLower;
 import static io.github.azagniotov.stubby4j.yaml.ConfigurableYAMLProperty.BODY;
+import static java.util.Collections.list;
 
 public class StubRepository {
 
@@ -48,7 +54,26 @@ public class StubRepository {
         this.matchedStubsCache = new ConcurrentHashMap<>();
     }
 
-    public StubResponse findStubResponseFor(final StubRequest incomingRequest) {
+    public StubRequest toStubRequest(final HttpServletRequest request) throws IOException {
+        final StubRequest.Builder builder = new StubRequest.Builder();
+        builder.withUrl(request.getPathInfo())
+                .withPost(extractPostRequestBody(request, "stubs"))
+                .withMethod(request.getMethod());
+
+        final Enumeration<String> headerNamesEnumeration = request.getHeaderNames();
+        final List<String> headerNames = isNotNull(headerNamesEnumeration) ? list(request.getHeaderNames()) : new LinkedList<>();
+        for (final String headerName : headerNames) {
+            final String headerValue = request.getHeader(headerName);
+
+            builder.withHeader(toLower(headerName), headerValue);
+        }
+
+        builder.withQuery(constructParamMap(request.getQueryString()));
+
+        return builder.build();
+    }
+
+    public StubResponse search(final StubRequest incomingRequest) {
         return findMatch(new StubHttpLifecycle.Builder().withRequest(incomingRequest).build());
     }
 
@@ -104,7 +129,7 @@ public class StubRepository {
      *
      * @param incomingStub {@link StubHttpLifecycle}
      * @return an {@link Optional} describing {@link StubHttpLifecycle} match, or an empty {@link Optional} if there was no match.
-     * @see StubRequest.Builder#fromHttpServletRequest(HttpServletRequest)
+     * @see #toStubRequest(HttpServletRequest)
      * @see StubHttpLifecycle#equals(Object)
      * @see StubRequest#equals(Object)
      * @see StubMatcher#matches(StubRequest, StubRequest)
